+++ fitty_frontend_app/lib/data/all_exercise_record.dart
