+++ fitty_frontend_app/lib/page/calendar_menu/tab/my_diet_calendar_tab.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:table_calendar/table_calendar.dart';
 
-class MyDietCalendar extends StatelessWidget {
-  const MyDietCalendar({super.key});
+class MyDietCalendarTab extends StatelessWidget {
+  const MyDietCalendarTab({super.key});
 
   @override
   Widget build(BuildContext context) {
