+++ fitty_frontend_app/lib/page/home_menu/tab/my_diet_home_tab.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class MyDietHome extends StatelessWidget {
-  const MyDietHome({
+class MyDietHomeTab extends StatelessWidget {
+  const MyDietHomeTab({
     super.key,
   });
