+++ fitty_frontend_app/lib/page/intro_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class IntroScreen extends StatelessWidget {
-  const IntroScreen({Key? key}) : super(key: key);
+class IntroPage extends StatelessWidget {
+  const IntroPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
