+++ fitty_frontend_app/lib/widget/exercise/exercise_widget.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 
 class ExerciseWidget extends StatelessWidget {
